backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
requests==2.31.0
openai==1.3.0
PyPDF2==3.0.1
python-docx==1.1.0
ezdxf==1.1.1
openpyxl==3.1.2
cryptography==41.0.7
bcrypt==4.1.2
passlib==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
bcrypt==4.1.2
passlib==1.7.4
python-multipart==0.0.6 